remove file extension; #87
